Added postman collection

+++ src/main/java/es/joaquin/ecommerce/infraestructure/CartItemRepositoryAdapter.java
@@ -2,7 +2,9 @@ package es.joaquin.ecommerce.infraestructure;
 
 import java.util.Optional;
 
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Component;
+import org.springframework.web.server.ResponseStatusException;
 
 import es.joaquin.ecommerce.domain.CartItemDto;
 import es.joaquin.ecommerce.domain.CartItemRepository;
@@ -33,12 +35,15 @@ public class CartItemRepositoryAdapter implements CartItemRepository{
 	public Optional<CartItemDto> addCartItem(Long cartId, Long prodId, Integer prodQuantity) {
 		Optional<ShoppingCartEntity> shoppingCartEntityOptional = shoppingCartJpaRepository.findById(cartId);
 		if (shoppingCartEntityOptional.isPresent()) {
-			Optional<ProductEntity> productEntityOptional = productJpaRepository.findById(prodId);
-			if (productEntityOptional.isPresent()) {
-				CartItemEntity cartItemEntity = new CartItemEntity(shoppingCartEntityOptional.get(), productEntityOptional.get(),prodQuantity);
-				cartItemJpaRepository.save(cartItemEntity);
-				return Optional.of(new CartItemDto(cartItemEntity.getId(), ProductRepositoryAdapter.toProductDto(cartItemEntity.getProduct()), cartItemEntity.getQuantity()));
+			if (!shoppingCartEntityOptional.get().getClosed()) {	
+				Optional<ProductEntity> productEntityOptional = productJpaRepository.findById(prodId);
+				if (productEntityOptional.isPresent()) {
+					CartItemEntity cartItemEntity = new CartItemEntity(shoppingCartEntityOptional.get(), productEntityOptional.get(),prodQuantity);
+					cartItemJpaRepository.save(cartItemEntity);
+					return Optional.of(new CartItemDto(cartItemEntity.getId(), ProductRepositoryAdapter.toProductDto(cartItemEntity.getProduct()), cartItemEntity.getQuantity()));
+				}
 			}
+			else throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Closed Shopping Carts can not be modified or deleted");
 		}
 		return Optional.empty();
 	}
@@ -47,11 +52,14 @@ public class CartItemRepositoryAdapter implements CartItemRepository{
 	public Boolean deleteItem(Long cartId, Long prodId) {
 		Optional<ShoppingCartEntity> shoppingCartEntityOptional = shoppingCartJpaRepository.findById(cartId);
 		if (shoppingCartEntityOptional.isPresent()) {
-			Optional<CartItemEntity> cartItemEntityOptional = cartItemJpaRepository.findById(prodId);
-			if (cartItemEntityOptional.isPresent()) {
-				cartItemJpaRepository.deleteById(prodId);
-				return true;
+			if (!shoppingCartEntityOptional.get().getClosed()) {
+				Optional<CartItemEntity> cartItemEntityOptional = cartItemJpaRepository.findById(prodId);
+				if (cartItemEntityOptional.isPresent()) {
+					cartItemJpaRepository.deleteById(prodId);
+					return true;
+				}
 			}
+			else throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Closed Shopping Carts can not be modified or deleted");
 		}
 		return false;
 	}
